Rename getUserId.tsx to getExperimentationId.ts
