+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/page.tsx
@@ -33,7 +33,7 @@ import { getEnabledFeatureFlags } from "../../../../../../../db/tables/featureFl
 import { getAttributionsFromCookiesOrDb } from "../../../../../../functions/server/attributions";
 import { checkSession } from "../../../../../../functions/server/checkSession";
 import { isPrePlusUser } from "../../../../../../functions/server/isPrePlusUser";
-import { getUserId } from "../../../../../../functions/server/getUserId";
+import { getUserId } from "../../../../../../functions/server/getExperimentationId";
 import { getElapsedTimeInDaysSinceInitialScan } from "../../../../../../functions/server/getElapsedTimeInDaysSinceInitialScan";
 import { getExperiments } from "../../../../../../functions/server/getExperiments";
 import { getLocale } from "../../../../../../functions/universal/getLocale";
